refactor: features 구조 개선
