+++ src/features/chat/ui/connect.vue
@@ -11,9 +11,10 @@
 </template>
 
 <script setup lang="ts">
-import { storeToRefs } from "pinia";
+import { watch } from "vue";
 import { watchEffect } from "vue";
 
+import { Room, User } from "@/entities/chat/model";
 import {
   connected,
   connect_failed,
@@ -24,15 +25,31 @@ import {
 } from "../service/event_helper";
 import useChatStore from "../store/useChatStore";
 
+const { connecting, room, current_user } = defineProps({
+  connecting: Boolean,
+  room: Room,
+  current_user: User,
+});
 const { insert_message, alarm_typing } = useChatStore();
-const { connecting } = storeToRefs(useChatStore());
+const store = useChatStore();
+
+// store에 props를 업데이트
+watch(
+  () => ({ connecting, room, current_user }),
+  (props) => {
+    store.connecting = props.connecting;
+    store.room = props.room;
+    store.current_user = props.current_user!;
+  },
+  { immediate: true, deep: true },
+);
 
 watchEffect(() => {
-  if (!!connecting.value) {
+  if (!!store.connecting) {
     // 소켓 이벤트 리스너 등록
-    connected(() => (connecting.value = true)); // 연결 성공
-    connect_failed(() => (connecting.value = false)); // 연결 실패
-    disconnected(() => (connecting.value = false)); // 연결 종료
+    connected(() => (store.connecting = true)); // 연결 성공
+    connect_failed(() => (store.connecting = false)); // 연결 실패
+    disconnected(() => (store.connecting = false)); // 연결 종료
     message_received(insert_message); // 일반 메시지 수신
     system_message_received(insert_message); // 시스템 메시지 수신
     handle_typing_message(alarm_typing); // 타이핑 알림
