+++ src/features/room/index.ts
@@ -1,0 +1,2 @@
+export { default as useRoomStore } from "./store/useRoomStore";
+export * from "./ui";
