+++ src/pages/index.vue
@@ -10,14 +10,16 @@
       <template #side-header> <make-room /> </template>
       <template #side> <room-list /> </template>
 
-      <template #content>
-        <Chat
+      <template #connect>
+        <connect
           v-if="!!selected_room?.id"
           :connecting="connecting"
           :room="selected_room"
           :current_user="current_user"
         />
       </template>
+
+      <template #content> <content /> </template>
 
       <template #toolbar> <actions /> </template>
 
@@ -29,10 +31,8 @@
 <script setup>
 import { storeToRefs } from "pinia";
 
-import useRoomStore from "@/features/room/store/useRoomStore";
-import { Register, MakeRoom, RoomList } from "@/features/room/ui";
-import { Submit, Actions, Title, Search } from "@/features/chat/ui";
-import Chat from "@/features/chat";
+import { Register, MakeRoom, RoomList, useRoomStore } from "@/features/room";
+import { Connect, Content, Submit, Actions, Title, Search } from "@/features/chat";
 import Layout from "./ui/layout.vue";
 
 // feature간 데이터 공유
